Create README.md
This is an academic project I did on Object-Oriented Programming course.
In this project, I was responsible for writing a C/S structured facility. The client can upload file to a server, download file from a server, browse file system(permission check applied) of a server, search for files in a directory and its sub-directories with file name that contains a pattern(e.g: *.cpp) on the server and search for files that contains a text pattern(e.g: int main) on the server. The connection is implemented with socket and asynchronous communication. The client UI is implemented in WPF.

Please be sure that Visual Studio 2013 or newer is installed on your PC.
Run compile.bat to compile the project
Run run.bat to run the project.